Moved PendingMessage class to its own file
