+++ RiptideNetworking/RiptideNetworking/Transports/RudpTransport/RudpClient.cs
@@ -245,7 +245,7 @@ namespace RiptideNetworking.Transports.RudpTransport
 
             lock (peer.PendingMessages)
             {
-                foreach (RudpPeer.PendingMessage pendingMessage in peer.PendingMessages.Values)
+                foreach (PendingMessage pendingMessage in peer.PendingMessages.Values)
                     pendingMessage.Clear(false);
 
                 peer.PendingMessages.Clear();
