+++ RiptideNetworking/RiptideNetworking/Transports/RudpTransport/RudpConnection.cs
@@ -65,7 +65,7 @@ namespace RiptideNetworking.Transports.RudpTransport
 
             lock (Peer.PendingMessages)
             {
-                foreach (RudpPeer.PendingMessage pendingMessage in Peer.PendingMessages.Values)
+                foreach (PendingMessage pendingMessage in Peer.PendingMessages.Values)
                     pendingMessage.Clear(false);
 
                 Peer.PendingMessages.Clear();
